feat(HighScores): embellish the New High Score View
* Animate the cards to be Faced Up/Down every seconds
* Add a background to the screen as well
* Fix the padding of the generated image
