+++ Shared/HighScores/NewHighScoreView.swift
@@ -7,8 +7,9 @@ struct NewHighScoreView: View {
 
     var body: some View {
         WithViewStore(store) { viewStore in
-            Form {
-                cardsAnimation
+            VStack {
+                Spacer()
+                VictoryCardsView()
                 TextField("Your name", text: $name, onCommit: {
                     submit(with: viewStore)
                 })
@@ -18,7 +19,14 @@ struct NewHighScoreView: View {
                 Button(action: { submit(with: viewStore) }, label: {
                     Text("Add my new high score")
                 })
+                Spacer()
             }
+            .background(
+                Image("Motif")
+                    .resizable(resizingMode: .tile)
+                    .renderingMode(.template)
+                    .opacity(1/10)
+            )
         }
     }
 
@@ -35,84 +43,6 @@ struct NewHighScoreView: View {
             )))
         }
         viewStore.send(.newHighScoreEntered)
-    }
-
-    private var cardsAnimation: some View {
-        HStack {
-            Spacer()
-            ZStack {
-                ForEach(0..<5) { cardNumber in
-                    CCardView(color: randomColor, image: randomArt, isFacedUp: .constant(true))
-                        .frame(width: 60, height: 60)
-                        .rotationEffect(angle(cardNumber: cardNumber))
-                        .offset(offset(cardNumber: cardNumber))
-                }
-                .padding()
-            }
-            Spacer()
-        }
-    }
-
-    private func angle(cardNumber: Int) -> Angle {
-        .radians(-.pi/16 * 4 + 2 * .pi/16 * Double(cardNumber))
-    }
-
-    private func offset(cardNumber: Int) -> CGSize {
-        CGSize(
-            width: -16 * 4 + 2 * 16 * cardNumber,
-            height: 4 * (cardNumber * cardNumber) - 16 * cardNumber
-        )
-    }
-
-    private var randomColor: Color {
-        [Color.green, Color.blue, Color.red].randomElement() ?? .red
-    }
-
-    private var randomArt: Image {
-        Art.allCases.randomElement()?.image ?? Art.artDeco.image
-    }
-}
-
-struct CCardView: View {
-    let color: Color
-    let image: Image
-    @Binding var isFacedUp: Bool
-    let action: () -> Void = { }
-    private static let turnCardAnimationDuration: Double = 2/5
-
-    var body: some View {
-        ZStack {
-            if !isFacedUp {
-                Button(action: action) {
-                    color.transition(turnTransition)
-                }
-                .buttonStyle(PlainButtonStyle())
-            } else {
-                image
-                    .renderingMode(.original)
-                    .resizable()
-                    .font(.largeTitle)
-                    .transition(turnTransition)
-            }
-        }
-        .modifier(AddCardStyle(foregroundColor: color))
-        .rotation3DEffect(
-            isFacedUp
-                ? .radians(.pi)
-                : .zero,
-            axis: (x: 0.0, y: 1.0, z: 0.0),
-            perspective: 1/3
-        )
-        .animation(.easeInOut(duration: Self.turnCardAnimationDuration))
-        .rotation3DEffect(.radians(.pi), axis: (x: 0.0, y: 1.0, z: 0.0))
-    }
-
-    private var turnTransition: AnyTransition {
-        AnyTransition.opacity.animation(
-            Animation
-                .linear(duration: 0.01)
-                .delay(Self.turnCardAnimationDuration/2)
-        )
     }
 }
